mini refactoring
